+++ src/Lib/ICS.php
@@ -598,6 +598,12 @@ class ICS {
 	 * Action effectuée par le Courrielleur : CREATE, MOVE, DELETE, COPY
 	 */
 	const X_CM2V3_ACTION = 'X-CM2V3-ACTION';
+	
+	// M2
+	/**
+	 * Gestion du champ X-M2-ORG-MAIL pour l'ORGANIZER
+	 */
+	const X_M2_ORG_MAIL = 'X-M2-ORG-MAIL';
 
 	// CALDAV X
 	/**